refactor: change `database_insert_batch_size` type from `u64` to `usize` across extractor, CLI, and related modules

+++ tycho-indexer/src/main.rs
@@ -421,7 +421,7 @@ async fn build_all_extractors(
     s3_bucket: Option<&str>,
     substreams_api_token: &str,
     cached_gw: &CachedGateway,
-    database_insert_batch_size: u64,
+    database_insert_batch_size: usize,
     token_pre_processor: &EthereumTokenPreProcessor,
     rpc_url: &str,
     runtime: Option<&tokio::runtime::Handle>,
